Update CustomAPICore.js
Code cleaned up

Aspects of the day now lock, so only one user can win them. Anyone who rolls the same value will be pushed +1 or -1 away from the winning value or item.

Adding items to "Aspect Of The Day" is now much easier. 
The readme will be updated soon with the updated guide! 

added a "const specialInteractions." 
which allows you to decide custom messages for interactions! 

and much more!